This is the joystick Library
When called the read channel will return the analog value on that channel. The channels that can be returned are (vrx_channel) or (vry_channel) and (swt_channel). vrx_channel is the channel that checks the position of the Joystick on the x axis. vry_channel does the same thing but checks the positon of the Joysick on the y axis. The swt_channel is the button on the joystick that is pressed when you "click" the joystick. *Note* this function will not work unless it is run on my pi so we will have to meet in person to figure out the range of motion that we will need to map to move the sprite in the corresponding directions.